refactor(API): Extracts postOAuth2TokenRequest to request handlers.
