+++ api/src/controllers/AuthController.js
@@ -1,13 +1,6 @@
-import passport from 'passport';
 import jsonwebtoken from 'jsonwebtoken';
-import { v4 as uuid } from 'uuid';
 import User from 'lib/models/user';
-import OAuthToken from 'lib/models/oAuthToken';
 import { AUTH_JWT_SUCCESS } from 'lib/constants/routes';
-import {
-  ACCESS_TOKEN_VALIDITY_PERIOD_SEC,
-  DEFAULT_PASSPORT_OPTIONS
-} from 'lib/constants/auth';
 import Unauthorized from 'lib/errors/Unauthorised';
 import { createOrgJWT, createUserJWT } from 'api/auth/jwt';
 import catchErrors from 'api/controllers/utils/catchErrors';
@@ -67,49 +60,9 @@ const success = (req, res) => {
   res.send('Login success!');
 };
 
-const issueOAuth2AccessToken = (req, res) => {
-  passport.authenticate('OAuth2_Authorization', DEFAULT_PASSPORT_OPTIONS, (err, client) => {
-    if (err) {
-      if (err.isClientError) {
-        res.status(400);
-        res.send({ error: err.error });
-        return;
-      }
-      res.status(500);
-      res.send(err.error);
-      return;
-    }
-
-    const accessToken = uuid();
-    const createdAt = new Date();
-    const expireAt = new Date(createdAt.getTime());
-    expireAt.setSeconds(createdAt.getSeconds() + ACCESS_TOKEN_VALIDITY_PERIOD_SEC);
-
-    OAuthToken.create({
-      clientId: client._id,
-      accessToken,
-      createdAt,
-      expireAt,
-    }, (err) => {
-      if (err) {
-        res.status(500);
-        res.send(err);
-        return;
-      }
-      res.status(200);
-      res.send({
-        access_token: accessToken,
-        token_type: 'bearer',
-        expires_in: ACCESS_TOKEN_VALIDITY_PERIOD_SEC,
-      });
-    });
-  })(req, res);
-};
-
 export default {
   clientInfo,
   jwtRefresh,
   googleSuccess,
   success,
-  issueOAuth2AccessToken,
 };
